LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/valueMode-expected.txt


FAIL value IDL attribute of input type hidden without value attribute assert_equals: expected "" but got "foo"
FAIL value IDL attribute of input type hidden with value attribute assert_equals: expected "bar" but got "foo"
PASS value IDL attribute of input type text without value attribute
PASS value IDL attribute of input type text with value attribute
PASS value IDL attribute of input type search without value attribute
PASS value IDL attribute of input type search with value attribute
PASS value IDL attribute of input type tel without value attribute
PASS value IDL attribute of input type tel with value attribute
PASS value IDL attribute of input type url without value attribute
PASS value IDL attribute of input type url with value attribute
PASS value IDL attribute of input type email without value attribute
PASS value IDL attribute of input type email with value attribute
PASS value IDL attribute of input type password without value attribute
PASS value IDL attribute of input type password with value attribute
FAIL value IDL attribute of input type datetime without value attribute assert_equals: expected "" but got "foo"
FAIL value IDL attribute of input type datetime with value attribute assert_equals: expected "" but got "foo"
PASS value IDL attribute of input type date without value attribute
PASS value IDL attribute of input type date with value attribute
PASS value IDL attribute of input type month without value attribute
PASS value IDL attribute of input type month with value attribute
PASS value IDL attribute of input type week without value attribute
PASS value IDL attribute of input type week with value attribute
PASS value IDL attribute of input type time without value attribute
PASS value IDL attribute of input type time with value attribute
PASS value IDL attribute of input type number without value attribute
PASS value IDL attribute of input type number with value attribute
PASS value IDL attribute of input type range without value attribute
PASS value IDL attribute of input type range with value attribute
FAIL value IDL attribute of input type color without value attribute assert_equals: expected "#000000" but got "foo"
FAIL value IDL attribute of input type color with value attribute assert_equals: expected "#000000" but got "foo"
FAIL value IDL attribute of input type checkbox without value attribute assert_equals: expected "on" but got "foo"
FAIL value IDL attribute of input type checkbox with value attribute assert_equals: expected "bar" but got "foo"
FAIL value IDL attribute of input type radio without value attribute assert_equals: expected "on" but got "foo"
FAIL value IDL attribute of input type radio with value attribute assert_equals: expected "bar" but got "foo"
FAIL value IDL attribute of input type submit without value attribute assert_equals: expected "" but got "foo"
FAIL value IDL attribute of input type submit with value attribute assert_equals: expected "bar" but got "foo"
FAIL value IDL attribute of input type image without value attribute assert_equals: expected "" but got "foo"
FAIL value IDL attribute of input type image with value attribute assert_equals: expected "bar" but got "foo"
FAIL value IDL attribute of input type reset without value attribute assert_equals: expected "" but got "foo"
FAIL value IDL attribute of input type reset with value attribute assert_equals: expected "bar" but got "foo"
FAIL value IDL attribute of input type button without value attribute assert_equals: expected "" but got "foo"
FAIL value IDL attribute of input type button with value attribute assert_equals: expected "bar" but got "foo"
